Split enums to own files
Makes it easier to find and not create dupes...
