+++ GDWeave/Godot/Token.cs
@@ -1,4 +1,4 @@
-﻿// TODO: engine-dependent
+// TODO: engine-dependent
 
 using System.Diagnostics.CodeAnalysis;
 
@@ -55,103 +55,3 @@ public class IdentifierToken : Token {
     }
 }
 
-public enum TokenType {
-    Empty,
-    Identifier,
-    Constant,
-    Self,
-    BuiltInType,
-    BuiltInFunc,
-    OpIn,
-    OpEqual,
-    OpNotEqual,
-    OpLess,
-    OpLessEqual,
-    OpGreater,
-    OpGreaterEqual,
-    OpAnd,
-    OpOr,
-    OpNot,
-    OpAdd,
-    OpSub,
-    OpMul,
-    OpDiv,
-    OpMod,
-    OpShiftLeft,
-    OpShiftRight,
-    OpAssign,
-    OpAssignAdd,
-    OpAssignSub,
-    OpAssignMul,
-    OpAssignDiv,
-    OpAssignMod,
-    OpAssignShiftLeft,
-    OpAssignShiftRight,
-    OpAssignBitAnd,
-    OpAssignBitOr,
-    OpAssignBitXor,
-    OpBitAnd,
-    OpBitOr,
-    OpBitXor,
-    OpBitInvert,
-    CfIf,
-    CfElif,
-    CfElse,
-    CfFor,
-    CfWhile,
-    CfBreak,
-    CfContinue,
-    CfPass,
-    CfReturn,
-    CfMatch,
-    PrFunction,
-    PrClass,
-    PrClassName,
-    PrExtends,
-    PrIs,
-    PrOnready,
-    PrTool,
-    PrStatic,
-    PrExport,
-    PrSetget,
-    PrConst,
-    PrVar,
-    PrAs,
-    PrVoid,
-    PrEnum,
-    PrPreload,
-    PrAssert,
-    PrYield,
-    PrSignal,
-    PrBreakpoint,
-    PrRemote,
-    PrSync,
-    PrMaster,
-    PrSlave,
-    PrPuppet,
-    PrRemotesync,
-    PrMastersync,
-    PrPuppetsync,
-    BracketOpen,
-    BracketClose,
-    CurlyBracketOpen,
-    CurlyBracketClose,
-    ParenthesisOpen,
-    ParenthesisClose,
-    Comma,
-    Semicolon,
-    Period,
-    QuestionMark,
-    Colon,
-    Dollar,
-    ForwardArrow,
-    Newline,
-    ConstPi,
-    ConstTau,
-    Wildcard,
-    ConstInf,
-    ConstNan,
-    Error,
-    Eof,
-    Cursor
-}
