+++ GDWeave/Godot/Variant.cs
@@ -9,40 +9,3 @@ public abstract class Variant : IEquatable<Variant>, ICloneable {
     public abstract object GetValue();
 }
 
-public enum VariantType : uint {
-    Nil,
-
-    // atomic types
-    Bool,
-    Int,
-    Real,
-    String,
-
-    // math types
-    Vector2,
-    Rect2,
-    Vector3,
-    Transform2D,
-    Plane,
-    Quat,
-    AaBb,
-    Basis,
-    Transform,
-
-    // misc types
-    Color,
-    NodePath,
-    Rid,
-    Object,
-    Dictionary,
-    Array,
-
-    // arrays
-    PoolByteArray,
-    PoolIntArray,
-    PoolRealArray,
-    PoolStringArray,
-    PoolVector2Array,
-    PoolVector3Array,
-    PoolColorArray
-}
